requirements.txt
asgiref==3.7.2
Django==5.0.1
django-prometheus==2.3.1
djangorestframework==3.14.0
prometheus-client==0.19.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
